created functions to add single date data to the chain
