+++ src/App.jsx
@@ -1,7 +1,7 @@
 import { Route, Routes } from 'react-router-dom'
 
 import { Navbar } from './components'
-import { Home, User } from './pages'
+import { Home, User, Coach } from './pages'
 
 const App = () => {
   return (
@@ -10,6 +10,7 @@ const App = () => {
       <Routes>
         <Route path="/" element={ <Home /> } />
         <Route path="/user" element={ <User /> } />
+        <Route path="/coach" element={ <Coach /> } />
       </Routes>
     </div>
   )
