+++ src/pages/Coach.jsx
@@ -1,0 +1,9 @@
+import './coach.css'
+
+const Coach = () => {
+  return (
+    <div>Coach</div>
+  )
+}
+
+export default Coach
