+++ src/pages/index.js
@@ -1,2 +1,3 @@
 export { default as Home } from './Home'
 export { default as User } from './User'
+export { default as Coach } from './Coach'
